rnn_predictor/requirements.txt
numpy
matplotlib
tensorflow
pyspellchecker
